+++ docker-compose.yml
@@ -144,9 +144,13 @@ services:
         dockerfile: ./Dockerfile-dev
       volumes:
         - .:/app
+      deploy:
+        mode: replicated
+        replicas: 4
+      restart: unless-stopped
       networks:
         testing_net:
-          ipv4_address: 172.28.1.16
+#          ipv4_address: 172.28.1.16
 
 
     telemetry-backend-core:
