+++ package.json
@@ -6,8 +6,8 @@
     "packages/common",
     "packages/gateway",
     "packages/telemetry",
-    "packages/core",
-    "packages/worker"
+    "packages/worker",
+    "packages/core"
   ],
   "scripts": {
     "create-config:all": "yarn workspaces foreach run create-config",
