+++ packages/common/src/index.ts
@@ -6,5 +6,18 @@ import ApiHandler from "./ApiHandler";
 import * as Constants from "./constants";
 import * as Types from "./types";
 import * as Util from "./util";
+import * as Constraints from "./constraints";
+import * as Score from "./score";
 
-export { ApiHandler, ChainData, Constants, Db, Config, logger, Types, Util };
+export {
+  ApiHandler,
+  ChainData,
+  Constants,
+  Db,
+  Config,
+  logger,
+  Types,
+  Util,
+  Constraints,
+  Score,
+};
