+++ packages/worker/package.json
@@ -36,8 +36,7 @@
     "@typescript-eslint/parser": "^5.40.0",
     "ava": "^4.3.3",
     "concurrently": "^7.4.0",
-    "eslint": "^8.25.0",
-    "eslint-plugin-security": "^1.5.0",
+    "eslint": "8.22.0",
     "nodemon": "^2.0.20"
   },
   "ava": {
