+++ packages/worker/src/jobs/index.ts
@@ -1,1 +1,13 @@
 export * from "./ReleaseMonitorJob";
+export * from "./ConstraintsJob";
+export * from "./EraStatsJob";
+export * from "./EraPointsJob";
+export * from "./ValidatorPrefJob";
+export * from "./SessionKeyJob";
+export * from "./InclusionJob";
+export * from "./ActiveValidatorJob";
+export * from "./LocationStatsJob";
+export * from "./CouncilJob";
+export * from "./DemocracyJob";
+export * from "./NominatorJob";
+export * from "./DelegationJob";
