+++ packages/worker/src/workers/index.ts
@@ -1,1 +1,3 @@
 export * from "./ReleaseMonitorWorker";
+export * from "./ConstraintsWorker";
+export * from "./EraStatsWorker";
